back-end/sunway_cityhack/requirements.txt
alabaster==0.7.10
args==0.1.0
astroid==1.5.3
Babel==2.4.0
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
clint==0.5.1
configparser==3.5.0
coreapi==2.3.1
coreschema==0.0.4
cycler==0.10.0
Cython==0.27.2
darkflow==1.0.0
decorator==4.1.1
Django==1.11.4
django-after-response==0.2.2
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
docutils==0.13.1
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
grequests==0.3.0
h5py==2.7.0
html5lib==0.9999999
idna==2.6
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.6
lazy-object-proxy==1.3.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
numpy==1.13.1
numpydoc==0.7.0
olefile==0.44
openapi-codec==1.3.2
pandas==0.20.3
pandocfilters==1.4.1
pathlib2==2.3.0
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
pika==0.11.0
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.4.0
psutil==5.2.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pydot==1.2.3
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.2
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
redis==2.10.6
requests==2.18.4
rope==0.10.5
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.0
subprocess32==3.2.7
tensorflow==1.4.0rc0
tensorflow-tensorboard==0.4.0rc1
testpath==0.3.1
Theano==0.9.0
tornado==4.5.1
tqdm==4.15.0
traitlets==4.3.2
typing==3.6.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
wincertstore==0.2
wrapt==1.10.10
ws4py==0.4.2
zope.dottedname==4.2
